+++ packages/react-native-the-sheet/src/bottom-sheet/index.ts
@@ -1,4 +1,5 @@
 export * from './bottom-sheet-flat-list'
+export * from './bottom-sheet-footer'
 export * from './bottom-sheet-handle'
 export * from './bottom-sheet-scroll-view'
 export * from './bottom-sheet-view'