+++ packages/react-native-embedded-stack-navigator/CHANGELOG.md
@@ -1,5 +1,19 @@
 # react-native-embedded-stack-navigator
 
+## 1.0.3
+
+### Patch Changes
+
+- Update `Backdrop` to use `Animated.View` to allow for animated styling
+
+- Add `BottomSheetFooter` component, which sticks to the bottom of the sheet
+
+- Add `sheetVisibleHeight` and `sheetVisibleRatio` to `BottomSheetContext`
+
+- Remove `useBottomSheetPositionTracker` hook, as you can now access the visible height and ratio directly from the context (⚠️ **BREAKING**)
+
+- Add `trackBottomSheetVisibleRatio` to `BottomSheetPositionTracker`
+
 ## 1.0.2
 
 ### Patch Changes
@@ -10,7 +24,7 @@
 
 ### Patch Changes
 
-- Rename embedded stack navigator
+- Rename embedded stack navigator (⚠️ **BREAKING**)
   - MiniStackRoute -> EmbeddedStackRoute
   - MiniStackNavigationApi -> EmbeddedStackNavigationApi
   - MiniStackNavigationContext -> EmbeddedStackNavigationContext